New IP address class in preparation for conversion to Postgres
The PostgreSQL ActiveRecord driver can handle the `inet` and `cidr`
types that Postgres supports, but it converts them to Ruby IPAddr
objects. In the past, we've been extending the IPAddr (and more recently
the IPAddress) class with some custom methods of our own. To support
Postgres fully, I'm now monkey-patching Ruby's IPAddr class when ANTFARM
is loaded.

This commit also includes a test case for the monkey-patching.