Returning true in `set_attributes_from_store` method
This tripped me up for a friggin' day or more! Since this method is used
as a before_validation callback, it needs to return true or ActiveRecord
thinks there was a validation error and doesn't save the model. It took
me a while to figure out because no error message was getting set! :-(